refactor: separately load slideshow page
